Rename composable
